1.1 initial cleanup
- Deleted obsolete files Flowchart.java and GenGUI.java
- Moved all classes from Gengen_V2.gengenv2 package to gengenv2, removing an unnecessary extra layer of addressing
